requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
djangorestframework-yaml==2.0.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==23.0
Pillow==9.4.0
pytz==2022.7
PyYAML==6.0
redis==4.4.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
validate-docbr==1.10.0
